[추가] 수익금으로 변경 및 시가컬럼 추가

+++ src/main/java/com/AiFunding/ToBi/entity/AccountStockDetailEntity.java
@@ -38,7 +38,7 @@ public class AccountStockDetailEntity implements Serializable {
 
     @NotNull
     @Column(name = "income")
-    private Double income;
+    private Integer income;
 
     @CreatedDate
     @Column(name = "create_at")
